+++ packages/article-main-standard/__tests__/shared.web.js
@@ -1,3 +1,4 @@
+import React from "react";
 import TestRenderer from "react-test-renderer";
 import {
   addSerializers,
@@ -7,8 +8,12 @@ import {
   minimalWebTransform,
   print
 } from "@times-components/jest-serializer";
-import "./mocks.web";
+import { UserState } from "./mocks.web";
 import shared from "./shared.base";
+import ArticleMainStandard from "../src/article-main-standard";
+import articleFixture, { testFixture } from "../fixtures/full-article";
+import { adConfig } from "./ad-mock";
+import articleProps from "./shared-article-props";
 
 const omitProps = new Set([
   "className",
@@ -29,17 +34,34 @@ export default () => {
   );
 
   beforeEach(() => {
-    const nuk = {
-      user: {
-        isLoggedIn: false
+    UserState.mockStates = [];
+  });
+
+  shared(TestRenderer.create, [
+    {
+      name: "should show topics when logged in or shared",
+      test() {
+        UserState.mockStates = [UserState.loggedInOrShared];
+
+        const output = TestRenderer.create(
+          <ArticleMainStandard
+            {...articleProps}
+            adConfig={adConfig}
+            analyticsStream={() => {}}
+            article={articleFixture(testFixture)}
+            onAuthorPress={() => {}}
+            onCommentGuidelinesPress={() => {}}
+            onCommentsPress={() => {}}
+            onLinkPress={() => {}}
+            onRelatedArticlePress={() => {}}
+            onTopicPress={() => {}}
+            onTwitterLinkPress={() => {}}
+            onVideoPress={() => {}}
+          />
+        );
+
+        expect(output).toMatchSnapshot();
       }
-    };
-    global.nuk = nuk;
-  });
-
-  afterEach(() => {
-    global.nuk = {};
-  });
-
-  shared(TestRenderer.create);
+    }
+  ]);
 };