+++ packages/article-main-standard/src/article-topics/index.web.js
@@ -1,5 +1,6 @@
 import React from "react";
 import ArticleTopics from "@times-components/article-topics";
+import UserState from "@times-components/user-state";
 import PropTypes from "prop-types";
 import styles from "../styles/article-topics";
 import TopicsMetaContainer from "../styles/article-topics/responsive";
@@ -7,11 +8,13 @@ import TopicsMetaContainer from "../styles/article-topics/responsive";
 const ShowTopics = ({ topics }) => {
   if (topics && topics.length > 0) {
     return (
-      <TopicsMetaContainer>
-        <nav>
-          <ArticleTopics style={styles.topicsMetaContainer} topics={topics} />
-        </nav>
-      </TopicsMetaContainer>
+      <UserState state={UserState.loggedInOrShared}>
+        <TopicsMetaContainer>
+          <nav>
+            <ArticleTopics style={styles.topicsMetaContainer} topics={topics} />
+          </nav>
+        </TopicsMetaContainer>
+      </UserState>
     );
   }