+++ packages/save-and-share-bar/__tests__/mocks.js
@@ -1,3 +1,8 @@
+/* eslint-disable import/prefer-default-export */
+import { mockUserState } from "@times-components/user-state";
+
+export const UserState = mockUserState();
+
 jest.mock("../src/bar-item", () => "BarItem");
 jest.mock("@times-components/save-star-web", () => "SaveStarWeb");