+++ packages/save-and-share-bar/src/save-and-share-bar.js
@@ -1,30 +1,27 @@
+/* eslint-env browser */
 import React, { Component } from "react";
 import PropTypes from "prop-types";
-import { ActivityIndicator, View, Text, Clipboard } from "react-native";
+import { View, Clipboard } from "react-native";
 import {
-  IconEmail,
   IconFacebook,
   IconTwitter,
   IconCopyLink
 } from "@times-components/icons";
 import SaveStar from "@times-components/save-star-web";
+import UserState from "@times-components/user-state";
 import getTokenisedArticleUrlApi from "./get-tokenised-article-url-api";
 import withTrackEvents from "./tracking/with-track-events";
 import SharingApiUrls from "./constants";
 import styles from "./styles";
 import BarItem from "./bar-item";
+import EmailShare from "./email-share";
 
 /* eslint-disable jsx-a11y/anchor-is-valid */
 class SaveAndShareBar extends Component {
   constructor(props) {
     super(props);
     this.copyToClipboard = this.copyToClipboard.bind(this);
-    this.handleOnShareEmailPress = this.handleOnShareEmailPress.bind(this);
     this.onSaveButtonPress = this.onSaveButtonPress.bind(this);
-
-    this.state = {
-      isLoading: false
-    };
   }
 
   /* eslint-disable class-methods-use-this */
@@ -39,27 +36,6 @@ class SaveAndShareBar extends Component {
     onCopyLink();
   }
 
-  handleOnShareEmailPress() {
-    /* eslint-env browser */
-    const { articleHeadline, articleId, getTokenisedShareUrl } = this.props;
-    this.setState({ isLoading: true });
-
-    getTokenisedShareUrl(articleId)
-      .then(res => {
-        const { data } = res;
-        if (data) {
-          this.setState({ isLoading: false });
-          const { url } = data.article.tokenisedUrl;
-          const mailtoEmailUrl = `mailto:?subject=${articleHeadline} from The Times&body=I thought you would be interested in this story from The Times%0A%0A${articleHeadline}%0A%0A${url}`;
-          window.location.assign(mailtoEmailUrl);
-        }
-      })
-      .catch(error => {
-        this.setState({ isLoading: false });
-        console.error("Error in connecting to api", error);
-      });
-  }
-
   render() {
     const {
       articleId,
@@ -71,24 +47,16 @@ class SaveAndShareBar extends Component {
       saveApi
     } = this.props;
 
-    const { isLoading } = this.state;
-
     return (
       <View style={styles.container}>
         {sharingEnabled && (
           <View style={styles.rowItem}>
-            <Text style={styles.label}>Share</Text>
-            <BarItem onPress={this.handleOnShareEmailPress}>
-              {isLoading ? (
-                <ActivityIndicator size="small" style={styles.activityLoader} />
-              ) : (
-                <IconEmail
-                  fillColour="currentColor"
-                  height={styles.svgIcon.height}
-                  title="Share by email client"
-                />
-              )}
-            </BarItem>
+            <UserState
+              state={UserState.subscriber}
+              fallback={<EmailShare {...this.props} shouldTokenise={false} />}
+            >
+              <EmailShare {...this.props} shouldTokenise />
+            </UserState>
             <BarItem
               onPress={onShareOnTwitter}
               target="_blank"
@@ -124,18 +92,20 @@ class SaveAndShareBar extends Component {
             </BarItem>
           </View>
         )}
-        {savingEnabled && (
-          <View style={styles.rowItem}>
-            <SaveStar
-              colour={styles.svgIcon.save.strokeColour}
-              hoverColor={styles.svgIcon.hoverFillColour}
-              articleId={articleId}
-              saveApi={saveApi}
-              height={styles.svgIcon.star.height}
-              onSaveButtonPress={this.onSaveButtonPress}
-            />
-          </View>
-        )}
+        {savingEnabled ? (
+          <UserState state={UserState.loggedIn} serverRender={false}>
+            <View style={styles.rowItem}>
+              <SaveStar
+                colour={styles.svgIcon.save.strokeColour}
+                hoverColor={styles.svgIcon.hoverFillColour}
+                articleId={articleId}
+                saveApi={saveApi}
+                height={styles.svgIcon.star.height}
+                onSaveButtonPress={this.onSaveButtonPress}
+              />
+            </View>
+          </UserState>
+        ) : null}
       </View>
     );
   }
@@ -149,6 +119,7 @@ SaveAndShareBar.propTypes = {
   getTokenisedShareUrl: PropTypes.func,
   onCopyLink: PropTypes.func.isRequired,
   onShareOnFB: PropTypes.func,
+  onShareEmail: PropTypes.func,
   onShareOnTwitter: PropTypes.func,
   saveApi: PropTypes.shape({
     bookmark: PropTypes.func.isRequired,
@@ -163,6 +134,7 @@ SaveAndShareBar.propTypes = {
 SaveAndShareBar.defaultProps = {
   onShareOnFB: () => {},
   onShareOnTwitter: () => {},
+  onShareEmail: () => {},
   getTokenisedShareUrl: getTokenisedArticleUrlApi
 };