+++ packages/article-label/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.84](https://github.com/newsuk/times-components/compare/@times-components/article-label@2.3.83...@times-components/article-label@2.3.84) (2019-06-13)
+
+**Note:** Version bump only for package @times-components/article-label
+
+
+
+
+
 ## [2.3.83](https://github.com/newsuk/times-components/compare/@times-components/article-label@2.3.82...@times-components/article-label@2.3.83) (2019-06-07)
 
 **Note:** Version bump only for package @times-components/article-label