+++ packages/article-list/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.0.25](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.0.24...@times-components/article-list@9.0.25) (2019-06-13)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
 ## [9.0.24](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.0.23...@times-components/article-list@9.0.24) (2019-06-10)
 
 **Note:** Version bump only for package @times-components/article-list