+++ packages/gestures/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.34](https://github.com/newsuk/times-components/compare/@times-components/gestures@4.1.33...@times-components/gestures@4.1.34) (2019-06-13)
+
+**Note:** Version bump only for package @times-components/gestures
+
+
+
+
+
 ## [4.1.33](https://github.com/newsuk/times-components/compare/@times-components/gestures@4.1.32...@times-components/gestures@4.1.33) (2019-06-07)
 
 **Note:** Version bump only for package @times-components/gestures