+++ packages/pull-quote/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.92](https://github.com/newsuk/times-components/compare/@times-components/pull-quote@3.4.91...@times-components/pull-quote@3.4.92) (2019-06-13)
+
+**Note:** Version bump only for package @times-components/pull-quote
+
+
+
+
+
 ## [3.4.91](https://github.com/newsuk/times-components/compare/@times-components/pull-quote@3.4.90...@times-components/pull-quote@3.4.91) (2019-06-07)
 
 **Note:** Version bump only for package @times-components/pull-quote