+++ packages/save-star-web/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.8](https://github.com/newsuk/times-components/compare/@times-components/save-star-web@0.2.7...@times-components/save-star-web@0.2.8) (2019-06-13)
+
+**Note:** Version bump only for package @times-components/save-star-web
+
+
+
+
+
 ## [0.2.7](https://github.com/newsuk/times-components/compare/@times-components/save-star-web@0.2.6...@times-components/save-star-web@0.2.7) (2019-06-11)
 
 **Note:** Version bump only for package @times-components/save-star-web