+++ packages/slice-layout/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.29.8](https://github.com/newsuk/times-components/compare/@times-components/slice-layout@0.29.7...@times-components/slice-layout@0.29.8) (2019-06-13)
+
+**Note:** Version bump only for package @times-components/slice-layout
+
+
+
+
+
 ## [0.29.7](https://github.com/newsuk/times-components/compare/@times-components/slice-layout@0.29.6...@times-components/slice-layout@0.29.7) (2019-06-07)
 
 **Note:** Version bump only for package @times-components/slice-layout