+++ packages/sticky/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.5](https://github.com/newsuk/times-components/compare/@times-components/sticky@0.0.4...@times-components/sticky@0.0.5) (2019-06-13)
+
+**Note:** Version bump only for package @times-components/sticky
+
+
+
+
+
 ## [0.0.4](https://github.com/newsuk/times-components/compare/@times-components/sticky@0.0.3...@times-components/sticky@0.0.4) (2019-06-07)
 
 **Note:** Version bump only for package @times-components/sticky